+++ apps/infrastructure/CHANGELOG.md
@@ -1,5 +1,23 @@
 # infrastructure
 
+## 0.3.0
+
+### Minor Changes
+
+- b4f0221: Create DNS for CMS Media Library and link it to the CDN
+
+## 0.2.0
+
+### Minor Changes
+
+- 60e0be7: [DEV-1489] - Add Active Campaign records
+
+## 0.1.2
+
+### Patch Changes
+
+- fd08262: [DEV-1525] Enable auto_minor_version_upgrade to cms database
+
 ## 0.1.1
 
 ### Patch Changes